diarisation-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2
python-multipart==0.0.9
numpy<2.0.0
huggingface_hub<0.24.0
